Dry up emitNativeMaybe, and apply it everywhere.
Also, hej: now recursive maps can be compiled!  (Previously they were
stymied simply because the maybe emission hadn't been pasted in.)
