+++ schema/gen/go/genString.go
@@ -62,42 +62,7 @@ func (g stringGenerator) EmitNativeBuilder(w io.Writer) {
 }
 
 func (g stringGenerator) EmitNativeMaybe(w io.Writer) {
-	// REVIEW: can this be extracted to the mixins package?  it doesn't appear to vary for kind.
-	doTemplate(`
-		type _{{ .Type | TypeSymbol }}__Maybe struct {
-			m schema.Maybe
-			v {{if not (MaybeUsesPtr .Type) }}_{{end}}{{ .Type | TypeSymbol }}
-		}
-		type Maybe{{ .Type | TypeSymbol }} = *_{{ .Type | TypeSymbol }}__Maybe
-
-		func (m Maybe{{ .Type | TypeSymbol }}) IsNull() bool {
-			return m.m == schema.Maybe_Null
-		}
-		func (m Maybe{{ .Type | TypeSymbol }}) IsUndefined() bool {
-			return m.m == schema.Maybe_Absent
-		}
-		func (m Maybe{{ .Type | TypeSymbol }}) Exists() bool {
-			return m.m == schema.Maybe_Value
-		}
-		func (m Maybe{{ .Type | TypeSymbol }}) AsNode() ipld.Node {
-			switch m.m {
-				case schema.Maybe_Absent:
-					return ipld.Undef
-				case schema.Maybe_Null:
-					return ipld.Null
-				case schema.Maybe_Value:
-					return {{if not (MaybeUsesPtr .Type) }}&{{end}}m.v
-				default:
-					panic("unreachable")
-			}
-		}
-		func (m Maybe{{ .Type | TypeSymbol }}) Must() {{ .Type | TypeSymbol }} {
-			if !m.Exists() {
-				panic("unbox of a maybe rejected")
-			}
-			return {{if not (MaybeUsesPtr .Type) }}&{{end}}m.v
-		}
-	`, w, g.AdjCfg, g)
+	emitNativeMaybe(w, g.AdjCfg, g)
 }
 
 // --- type info --->
